requirements.txt
Flask==3.0.0
flask-cors==4.0.0
easyocr==1.7.1
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pdf2image==1.17.0
numpy==1.24.3
gunicorn==21.2.0
torch==2.1.0
torchvision==0.16.0
